Address review comments and update versions of dependencies
