+++ oak-segment-aws/pom.xml
@@ -33,7 +33,7 @@
     <name>Oak Segment AWS</name>
 
     <properties>
-        <aws.version>1.11.475</aws.version>
+        <aws.version>1.11.700</aws.version>
         <sqlite4java.version>1.0.392</sqlite4java.version>
     </properties>
 
@@ -273,7 +273,7 @@
         <dependency>
             <groupId>com.amazonaws</groupId>
             <artifactId>DynamoDBLocal</artifactId>
-            <version>1.11.86</version>
+            <version>1.12.0</version>
             <scope>test</scope>
         </dependency>
         <dependency>
