+++ oak-segment-aws/src/main/java/org/apache/jackrabbit/oak/segment/aws/AwsPersistence.java
@@ -18,7 +18,6 @@ package org.apache.jackrabbit.oak.segment.aws;
 
 import java.io.IOException;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.jackrabbit.oak.segment.spi.monitor.FileStoreMonitor;
 import org.apache.jackrabbit.oak.segment.spi.monitor.IOMonitor;
 import org.apache.jackrabbit.oak.segment.spi.monitor.RemoteStoreMonitor;
@@ -37,20 +36,8 @@ public class AwsPersistence implements SegmentNodeStorePersistence {
 
     protected final AwsContext awsContext;
 
-    private final String fileNameSuffix;
-
     public AwsPersistence(AwsContext awsContext) {
-        this(awsContext, null);
-    }
-
-    public AwsPersistence(AwsContext awsContext, String id) {
-        if (StringUtils.isNotBlank(id)) {
-            this.awsContext = awsContext.withDirectory(id);
-            this.fileNameSuffix = "." + id;
-        } else {
-            this.awsContext = awsContext;
-            this.fileNameSuffix = "";
-        }
+        this.awsContext = awsContext;
     }
 
     @Override
@@ -77,12 +64,12 @@ public class AwsPersistence implements SegmentNodeStorePersistence {
 
     @Override
     public JournalFile getJournalFile() {
-        return new AwsJournalFile(awsContext, "journal" + fileNameSuffix + ".log");
+        return new AwsJournalFile(awsContext, "journal.log");
     }
 
     @Override
     public GCJournalFile getGCJournalFile() throws IOException {
-        return new AwsGCJournalFile(awsContext, "gc" + fileNameSuffix + ".log");
+        return new AwsGCJournalFile(awsContext, "gc.log");
     }
 
     @Override
@@ -92,6 +79,6 @@ public class AwsPersistence implements SegmentNodeStorePersistence {
 
     @Override
     public RepositoryLock lockRepository() throws IOException {
-        return new AwsRepositoryLock(awsContext, "repo" + fileNameSuffix + ".lock").lock();
+        return new AwsRepositoryLock(awsContext, "repo.lock").lock();
     }
 }
